pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-ticket-triage"
version = "0.1.0"
description = "Ticket triage OpenEnv environment"
requires-python = ">=3.10"
dependencies = [
  "openenv-core[core]>=0.2.2",
  "fastapi==0.135.2",
  "uvicorn==0.42.0",
  "pydantic==2.12.5",
  "httpx==0.28.1",
  "openai==2.30.0",
  "python-dotenv==1.2.2",
  "PyYAML==6.0.2"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
  "pytest-cov>=4.0.0"
]

[project.scripts]
server = "app:main"

[tool.setuptools]
include-package-data = true

[tool.setuptools.packages.find]
where = ["."]
include = ["ticket_triage_env*"]
